Tests/LibWeb/Text/expected/css/fit-content-parsing-with-whitespace.txt

"fit-content(10%)" became "fit-content(10%)"
"fit-content(       10%)" became "fit-content(10%)"
"fit-content(10%  )" became "fit-content(10%)"
"fit-content(   10%  )" became "fit-content(10%)"
